requirements.txt
requests
python-decouple
pyTelegramBotAPI